fix: address code review issues in OTA upload implementation
Agent-Logs-Url: https://github.com/Deddu3D/Aquarium-Controller-ESP32P4/sessions/0e01297e-ddec-4701-b286-b388b4e18341

+++ main/web_server.c
@@ -1575,7 +1575,7 @@ static esp_err_t api_ota_upload_post_handler(httpd_req_t *req)
             break;
         }
         received += (size_t)r;
-        int pct = (int)((int64_t)received * 100 / (int64_t)total);
+        int pct = (int)(received * 100 / total);
         ota_update_set_progress(pct > 100 ? 100 : pct);
     }
